refactor: move `flow-types` back, add extra directory for ts tests.
