+++ package.json
@@ -17,7 +17,6 @@
   "main": "dist/immutable.js",
   "module": "dist/immutable.es.js",
   "typings": "dist/immutable-nonambient.d.ts",
-  "types": "type-definitions/ts-types",
   "typescript": {
     "definition": "dist/immutable.d.ts"
   },
@@ -37,8 +36,8 @@
     "testonly": "./resources/jest",
     "test": "run-s format build lint testonly test:types:*",
     "test:travis": "npm run test && ./resources/check-changes",
-    "test:types:ts": "dtslint type-definitions/ts-types/tests",
-    "test:types:flow": "flow check type-definitions/flow-types/tests",
+    "test:types:ts": "dtslint type-definitions/ts-tests",
+    "test:types:flow": "flow check type-definitions/tests",
     "perf": "node ./resources/bench.js",
     "start": "gulp --gulpfile gulpfile.js dev",
     "deploy": "(cd ./pages/out && git init && git config user.name \"Travis CI\" && git config user.email \"github@fb.com\" && git add . && git commit -m \"Deploy to GitHub Pages\" && git push --force --quiet \"https://${GH_TOKEN}@github.com/facebook/immutable-js.git\" master:gh-pages > /dev/null 2>1)"
@@ -62,7 +61,7 @@
     "browserify": "^5.11.2",
     "colors": "1.1.2",
     "del": "2.2.2",
-    "dtslint": "0.1.1",
+    "dtslint": "0.1.2",
     "eslint": "3.17.1",
     "eslint-config-airbnb": "14.1.0",
     "eslint-config-prettier": "1.5.0",
