+++ type-definitions/tests/.flowconfig
@@ -1,5 +1,5 @@
 [include]
-../../../
+../../
 
 [options]
 suppress_comment=\\(.\\|\n\\)*\\$ExpectError
