+++ type-definitions/ts-tests/index.d.ts
@@ -1,1 +1,2 @@
+// TypeScript Version: 2.2
 /* tslint:disable:no-useless-files */
