+++ type-definitions/ts-tests/tslint.json
@@ -4,6 +4,7 @@
     "no-var": false,
     "prefer-const": false,
     "no-useless-files": false,
+    "no-duplicate-imports": false,
     "no-relative-import-in-test": false
   }
 }
